split out stats
