+++ source/common/http/conn_manager_utility.h
@@ -7,8 +7,8 @@
 #include "envoy/network/connection.h"
 
 #include "common/http/conn_manager_impl.h"
-#include "common/http/http1/codec_impl.h"
-#include "common/http/http2/codec_impl.h"
+#include "common/http/http1/codec_stats.h"
+#include "common/http/http2/codec_stats.h"
 
 namespace Envoy {
 namespace Http {
