+++ source/common/http/http1/codec_impl.h
@@ -15,40 +15,16 @@
 #include "common/buffer/watermark_buffer.h"
 #include "common/common/assert.h"
 #include "common/common/statusor.h"
-#include "common/common/thread.h"
 #include "common/http/codec_helper.h"
 #include "common/http/codes.h"
 #include "common/http/header_map_impl.h"
+#include "common/http/http1/codec_stats.h"
 #include "common/http/http1/header_formatter.h"
 #include "common/http/status.h"
 
 namespace Envoy {
 namespace Http {
 namespace Http1 {
-
-/**
- * All stats for the HTTP/1 codec. @see stats_macros.h
- */
-#define ALL_HTTP1_CODEC_STATS(COUNTER)                                                             \
-  COUNTER(dropped_headers_with_underscores)                                                        \
-  COUNTER(metadata_not_supported_error)                                                            \
-  COUNTER(requests_rejected_with_underscores_in_headers)                                           \
-  COUNTER(response_flood)
-
-/**
- * Wrapper struct for the HTTP/1 codec stats. @see stats_macros.h
- */
-struct CodecStats {
-  using AtomicPtr = Thread::AtomicPtr<CodecStats, Thread::AtomicPtrAllocMode::DeleteOnDestruct>;
-
-  static CodecStats& atomicGet(AtomicPtr& ptr, Stats::Scope& scope) {
-    return *ptr.get([&scope]() -> CodecStats* {
-      return new CodecStats{ALL_HTTP1_CODEC_STATS(POOL_COUNTER_PREFIX(scope, "http1."))};
-    });
-  }
-
-  ALL_HTTP1_CODEC_STATS(GENERATE_COUNTER_STRUCT)
-};
 
 class ConnectionImpl;
