+++ source/common/upstream/upstream_impl.h
@@ -41,8 +41,8 @@
 #include "common/common/thread.h"
 #include "common/config/metadata.h"
 #include "common/config/well_known_names.h"
-#include "common/http/http1/codec_impl.h"
-#include "common/http/http2/codec_impl.h"
+#include "common/http/http1/codec_stats.h"
+#include "common/http/http2/codec_stats.h"
 #include "common/init/manager_impl.h"
 #include "common/network/utility.h"
 #include "common/shared_pool/shared_pool.h"
