+++ source/extensions/filters/network/http_connection_manager/config.h
@@ -18,8 +18,8 @@
 #include "common/common/logger.h"
 #include "common/http/conn_manager_impl.h"
 #include "common/http/date_provider_impl.h"
-#include "common/http/http1/codec_impl.h"
-#include "common/http/http2/codec_impl.h"
+#include "common/http/http1/codec_stats.h"
+#include "common/http/http2/codec_stats.h"
 #include "common/json/json_loader.h"
 #include "common/local_reply/local_reply.h"
 #include "common/router/rds_impl.h"
