+++ test/extensions/filters/listener/http_inspector/BUILD
@@ -18,6 +18,7 @@ envoy_extension_cc_test(
     tags = ["fails_on_windows"],
     deps = [
         "//source/common/common:hex_lib",
+        "//source/common/http:utility_lib",
         "//source/extensions/filters/listener/http_inspector:http_inspector_lib",
         "//test/mocks/api:api_mocks",
         "//test/mocks/network:network_mocks",
