+++ test/extensions/filters/listener/tls_inspector/BUILD
@@ -19,6 +19,7 @@ envoy_cc_test(
     srcs = ["tls_inspector_test.cc"],
     deps = [
         ":tls_utility_lib",
+        "//source/common/http:utility_lib",
         "//source/extensions/filters/listener/tls_inspector:config",
         "//source/extensions/filters/listener/tls_inspector:tls_inspector_lib",
         "//test/mocks/api:api_mocks",
@@ -37,6 +38,7 @@ envoy_extension_cc_benchmark_binary(
     ],
     deps = [
         ":tls_utility_lib",
+        "//source/common/http:utility_lib",
         "//source/common/network:listen_socket_lib",
         "//source/extensions/filters/listener/tls_inspector:tls_inspector_lib",
         "//test/mocks/api:api_mocks",
